chore: Update the vm's lalrpop to 0.14
