+++ parser/Cargo.toml
@@ -17,7 +17,7 @@ collect-mac = "0.1.0"
 env_logger = { version = "0.3.4", optional = true }
 itertools = "0.6.0"
 quick-error = "1.0.0"
-lalrpop-util = "0.13.1"
+lalrpop-util = "0.14.0"
 log = "0.3.6"
 pretty = "0.3.2"
 gluon_base = { path = "../base", version = "0.7.1" } # GLUON
